Updated models fields type of donation app

+++ donation/models.py
@@ -25,7 +25,7 @@ def section_position_default() -> int:
 class Section(models.Model): # type: ignore [django-manager-missing]
 	title = models.CharField(_('Заголовок'), max_length=255)
 	text = models.TextField(_('Текст'))
-	position = models.IntegerField(_('Позиция'), blank=True, default=section_position_default)
+	position = models.PositiveSmallIntegerField(_('Позиция'), blank=True, default=section_position_default)
 
 	class Meta(TypedModelMeta):
 		ordering = ('position',)
@@ -42,7 +42,7 @@ def button_position_default() -> int:
 class Button(models.Model): # type: ignore [django-manager-missing]
 	text = models.CharField(_('Текст'), max_length=255)
 	url = models.URLField(_('Ссылка'))
-	position = models.IntegerField(_('Позиция'), blank=True, default=button_position_default)
+	position = models.PositiveSmallIntegerField(_('Позиция'), blank=True, default=button_position_default)
 
 	class Meta(TypedModelMeta):
 		ordering = ('position',)
